New branch

+++ app/controllers/clients_controller.rb
@@ -7,7 +7,11 @@ class ClientsController < ApplicationController
 
     def show
         showClient = Client.find(params[:id])
-        render json: showClient
+        if showClient
+            render json: showClient
+        else
+            render json: {error: "Client not found" }, status: 404
+        end
     end
 
     def create
@@ -15,9 +19,18 @@ class ClientsController < ApplicationController
         render json: newClient
     end
 
+    def update
+        editClient = Client.find(params[:id])
+        render json: editClient.update(editClientParams)
+    end
+
     def destroy
         deleteClient = Client.find(params[:id])
-        deleteClient.destroy
+        if deleteClient
+            deleteClient.destroy
+        else
+            render json: { error: "Client not found"}, status: 404
+        end
     end
 
 
@@ -25,6 +38,10 @@ class ClientsController < ApplicationController
 
     def newClientParams
         params.permit( :first_name, :last_name, :phone_number, :email_address, :total_invoiced, :total_collected, :user_id)
+    end
+
+    def editClientParams
+        params.permit( :first_name, :last_name, :phone_number, :email_address)
     end
 
 end
